Consolidate support impersonation controller onto candidate controller
Also allow impersonating a candidate from the candidate show page.
