+++ app/views/support_interface/candidates/show.html.erb
@@ -4,6 +4,10 @@
   <%= @candidate.email_address %>
 <% end %>
 
+<% unless HostingEnvironment.production? %>
+  <%= button_to 'Sign in as this candidate', support_interface_impersonate_candidate_path(@candidate), class: 'govuk-button' %>
+<% end %>
+
 <% if @candidate.hide_in_reporting? %>
   <%= button_to 'Show in reporting', support_interface_show_candidate_path(@candidate), class: 'govuk-button govuk-button--secondary' %>
 <% else %>
